REFACTOR `CloudStorage` properties & methods related to local file deletion & upload streaming into standalone traits
